Move renamed buffs to TS on import
